ContactParser 3/src/contactParser.ts

class ContactParser {
  static parseContacts(contacts: string[]): Contact[] {
    const parsedContacts: Contact[] = [];

    for (const contact of contacts) {
      const parsedContact = this.parseContact(contact);
      parsedContacts.push(parsedContact);
    }

    return parsedContacts;
  }

  private static parseContact(contact: string): Contact {
    // Logic to parse the contact and extract relevant information
    // ...

    return {
      name,
      email,
      phone,
      address,
    };
  }
}

interface Contact {
  name: string;
  email: string;
  phone: string;
  address: string;
}

export default ContactParser;

